update poutine mechanism to retain compatibility and increase generality
