+++ pyro/poutine/lift_messenger.py
@@ -37,12 +37,14 @@ class LiftMessenger(Messenger):
         on the param names. If the param name does not match the
         name the keys in the prior, that param name is unchanged.
         """
+        # import pdb; pdb.set_trace()
         name = msg["name"]
         param_name = params.user_param_name(name)
         if isinstance(self.prior, dict):
             # prior is a dict of distributions
             if param_name in self.prior.keys():
                 msg["fn"] = self.prior[param_name]
+                msg["args"] = msg["args"][1:]
                 if isinstance(msg['fn'], Distribution):
                     msg["args"] = ()
                     msg["kwargs"] = {}
@@ -60,6 +62,7 @@ class LiftMessenger(Messenger):
                 # prior is a stochastic fn. block sample
                 msg["stop"] = True
             msg["fn"] = self.prior
+            msg["args"] = msg["args"][1:]
         else:
             # otherwise leave as is
             return None
